Reconnect to the same namenode after first write error
Summary:
Active namenode may close idle tcp connection, so next client write to
this connection will fail with "broken pipe" error.
But at this point there is no point in penalizing the namenode and try to
failover to other namenodes.
This change allows to try to reconnect to the same namenode after first write
error.

Reviewed By: himank

Signed-off-by: Evgeniy Firsov <firsov@uber.com>